Update ProductInstanceProps.tsx
In my opinion, the user should still be able to delete products from the contract when it is not yet signed. Even when I have sent a contract, changes might happen. These changes are not reflected in the pricing which makes life way harder as you have to calculate prices yourself and the product looks "cancelled".